requirements.txt
PyQt6
google-genai
pydantic
pyperclip
pywin32